+++ src/main/java/apptive/team1/friendly/common/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package apptive.team1.friendly.jwt;
+package apptive.team1.friendly.common.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
