+++ src/main/java/apptive/team1/friendly/common/jwt/JwtFilter.java
@@ -1,4 +1,4 @@
-package apptive.team1.friendly.jwt;
+package apptive.team1.friendly.common.jwt;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
